refactor: use interface for decoupling execution context
